src/alpesPartners/requirements.txt
flask[async]==3.1.1
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
pulsar-client==3.7.0
pulsar-client[avro]==3.7.0
PyDispatcher==2.0.6
SQLAlchemy==2.0.41
